+++ src/main/kotlin/com/github/xeonkryptos/eclipseprojectcreator/import/EclipseModuleImportListener.kt
@@ -5,7 +5,6 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.module.Module
 import com.intellij.openapi.project.ModuleListener
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.roots.ModifiableRootModel
 import com.intellij.openapi.roots.ModuleRootManager
 import org.jetbrains.idea.eclipse.EclipseXml
 
@@ -25,24 +24,15 @@ class EclipseModuleImportListener : ModuleListener {
             // The new module seems to be a new Eclipse project. Not yet sure, if it's an import of an existing module or a creation maybe with this plugin.
             if (relativeNioPath.toString() == EclipseXml.CLASSPATH_FILE) {
                 val modifiableRootModel = ModuleRootManager.getInstance(module).modifiableModel
-                val detectedIvyContainer = modifiableRootModel.orderEntries.any { orderEntry -> orderEntry.presentableName.startsWith(EclipseIvyCommons.IVYDE_CONTAINER_NAME) }
-                if (detectedIvyContainer) {
-                    removeIvyContainer(modifiableRootModel)
-                } else {
-                    modifiableRootModel.dispose()
-                }
-            }
-        }
-    }
-
-    private fun removeIvyContainer(modifiableRootModel: ModifiableRootModel) {
-        val application = ApplicationManager.getApplication()
-        application.invokeLater {
-            application.runWriteAction {
-                val ivyContainerOrderEntry = modifiableRootModel.orderEntries.firstOrNull { orderEntry -> orderEntry.presentableName.startsWith(EclipseIvyCommons.IVYDE_CONTAINER_NAME) }
-                if (ivyContainerOrderEntry != null) {
-                    modifiableRootModel.removeOrderEntry(ivyContainerOrderEntry)
-                    modifiableRootModel.commit()
+                val detectedIvyContainerEntry = modifiableRootModel.orderEntries.firstOrNull { orderEntry -> orderEntry.presentableName.startsWith(EclipseIvyCommons.IVYDE_CONTAINER_NAME) }
+                if (detectedIvyContainerEntry != null) {
+                    val application = ApplicationManager.getApplication()
+                    application.invokeLater {
+                        application.runWriteAction {
+                            modifiableRootModel.removeOrderEntry(detectedIvyContainerEntry)
+                            modifiableRootModel.commit()
+                        }
+                    }
                 } else {
                     modifiableRootModel.dispose()
                 }